Fix path got deleted in handling of move & close_write event
This is a best effort attempt to avoid the issue described in
abraunegg/onedrive#2588

The issue can be triggered if the following 3 events comes in a single
read() call for a speicfic file.  Note the missing IN_MOVED_TO event.

	IN_MOVED_FROM
	IN_CREATE
	IN_CLOSE_WRITE

The issue can be reproduced with the following python script operating
on the onedrive sync dir.

	import os

	os.rename('test.md', '/tmp/test.md')
	with open('test.md', 'w') as fout:
	    fout.write('hello')

While the change fixes this specific common case.  Working with
IN_MOVED_FROM and IN_MOVED_TO is known to be challenging ("Dealing with
rename() events" in inotify(7)).  In theory other scenarios can pop up
where the code does not handle it right.